Fix modules distribution
