+++ src/character/data/data.ts
@@ -1,4 +1,7 @@
-import { Fighter, King, Adviser, Squire } from "../types.js";
+import { Adviser } from "../adviser/types";
+import { Fighter } from "../fighter/types";
+import { King } from "../king/types";
+import { Squire } from "../squire/types";
 
 export const characters: (King | Adviser | Fighter | Squire)[] = [
   {
