+++ src/character/fighter/types.ts
@@ -1,0 +1,7 @@
+import { Character } from "../types";
+
+export type Fighter = Character & {
+  weapon: string;
+  dexterity: number;
+  phrase: "First I punch, then I ask";
+};
