+++ src/character/king/types.ts
@@ -1,0 +1,6 @@
+import { Character } from "../types";
+
+export type King = Character & {
+  yearsOfReign: number;
+  phrase: "You are all going to die";
+};
