+++ src/character/squire/types.ts
@@ -1,0 +1,7 @@
+import { Character } from "../types";
+
+export type Squire = Character & {
+  servesTo: string;
+  ballism: number;
+  phrase: "I am a looser";
+};
