+++ src/character/types.ts
@@ -1,0 +1,8 @@
+export type Character = {
+  name: string;
+  lastName: string;
+  age: number;
+  isAlive: boolean;
+  portraitUrl: string;
+  portraitDescription: string;
+};
